fix RPC communication failure
I couldn't figure out how to use `tDataqueueAdaptor` correctly. A single
dataqueue provides one data stream, and writing and then reading it
immediately would return the data that was just written. In theory,
there is no way to get it work correctly (except for peculiar
conditions).

This is why switched to `tMessageBufferCEP`, which seems to be working
well.